Revert "command create payment"
